requirements.txt
langchain==0.1.7
langchain_community==0.0.20
langchain_openai==0.0.6
python-dotenv==1.0.1
streamlit==1.31.0
